2.0.0
== 능력 개선 ==
휘트니
  - 휘트니의 스택형 효과가 WRECK의 영향에 받지 않던 현상이 수정됩니다.
  - 3스택 출혈 부여 시 3초의 쿨타임이 부여되며, 이는 WRECK의 영향에 받지 않도록 추가됩니다.
  - 4스택 흡혈량이 15.0%에서 12.5%로 조정됩니다.
  - 5스택 대미지 증가량이 3.0에서 1.5로 조정됩니다.
  - 6스택 대미지 감소량이 20.0%에서 15.0%로 조정됩니다.

코크스
  - 새로운 스킬 2개 [커스터마이징], [랜더마이즈]가 추가됩니다.
  - [이펙트 맛 좀 봐라]에서 한정적으로 등장하는 상태이상 <디버깅>이 추가됩니다.

듀얼
  - 사망 시 능력 명으로 배정받은 두 캐릭터의 능력명이 나오게 됩니다.
  - 설명과 달리 스턴과 봉인이 1초가 아닌 2초를 받던 현상이 수정됩니다.
  - 더이상 캐릭터에서 레인스타 애드온의 (유키) 능력이 나오지 않도록 수정됩니다.

에밀리
  - [알케미 에리어]의 상세 효과 내용이 [알케미 마스터리]의 효과가 바뀔 때 마다 채팅에 표기되도록 변경됩니다.
  - 능력 설명 중 [알케미 에리어]의 효과가 현 [알케미 마스터리]에 따라 변경됩니다.

아이리스
  - 래인보우의 쿨타임이 10초에서 5초로 감소됩니다.
  - 레인보우의 스킬 범위가 7블럭에서 10블럭으로 증가됩니다.
  - 악몽의 지속시간이 7초에서 4초로 감소됩니다.
  - 무지개 스택의 지속시간이 10초에서 20초로 증가되며, 컨피그에서 수정이 가능하도록 변경됩니다.
  - 레인보우와 나이트메어의 쿨타임이 WRECK 모드에서 최대 90%까지만 영향이 받도록 변경됩니다.

져스틴
  - 능력 설명 중 [두가지 인격]의 설명이 조금 간소화되었습니다.

팬텀 시프
  - 능력 발동 시 주변으로 파티클이 나오게 변경됩니다.
  - 믹스 능력자 전쟁에서 팬텀 시프의 능력 재설정이 재대로 이루어지지 않던 현상이 수정됩니다.
  - 분신이 사라진 후 자신의 갑옷이 타 플레이어에게는 보여지지 않던 현상이 수정됩니다.
  - 분신이 사망하면 그 분신이 장착하고 있던 아이템들이 드랍하던 현상이 수정됩니다.

리인카네이션
  - [환생]의 쿨타임이 설명란에 표기되지 않던 현상이 수정됩니다.
  - [환생] 중 무적 시간이 5초 이하에서 채팅에 도배되던 현상이 수정됩니다.

엑시즈
  - 리메이크 되었습니다. 게임 중 1회에 한하여 주변 플레이어를 자신에게 텔레포트 시키고, <저주>라는 특별한 디버프 효과를 줍니다.
  - <저주>는 해당 저주를 준 엑시즈 참여자가 사망하기 전까지 일정 주기로 관통대미지를 주는 디버프효과입니다.

== 시너지 개선 ==
베노미논, 베노미너거
  - 베노미논과 베노미너거의 맹독이 해당 능력자에게도 영향을 받던 현상이 수정됩니다.
  - 부활 시 자신이 부활한 사실이 채팅과 효과음으로 알 수 있게 됩니다.

디 엔드
  - 리메이크 되었습니다. 구조는 엑시즈와 동일하나 효과 대상자가 주변 플레이어가 아닌 모든 플레이어이며, 대미지를 주는 주기와 대미지량은 엑시즈보다 높게 설정되어있습니다.

== 룰렛 개선 ==
서프라이즈 발동 시 플레이어 수만큼 채팅이 뜨던 현상이 수정됩니다.
5초간 신속, 재생 1단계 부여 효과가 정상적으로 작동하지 않던 현상이 수정됩니다.
룰렛 효과가 추가됩니다.
  - 5초간 재생/힘/신속/저항/구속/나약함 중 하나 부여 (우선도 2)
  - 5초간 랜덤 플레이어의 근거리 공격 무시 (우선도 3)

+++ Addon/src/com/cokes86/cokesaddon/ability/list/Gambler.java
@@ -1,6 +1,7 @@
 package com.cokes86.cokesaddon.ability.list;
 
 import com.cokes86.cokesaddon.ability.CokesAbility;
+import com.cokes86.cokesaddon.ability.Config;
 import com.cokes86.cokesaddon.event.CEntityDamageEvent;
 import com.cokes86.cokesaddon.util.FunctionalInterfaces;
 import daybreak.abilitywar.ability.AbilityManifest;
@@ -60,7 +61,7 @@ public class Gambler extends CokesAbility implements ActiveHandler {
 	private boolean penaltyDice = false;
 	private Pair<Integer, Integer> penaltyDiceInitial = Pair.of(0, 0);
 
-	private final Cooldown cooldown = new Cooldown(PD_COOLDOWN.getValue().intValue());
+	private final Cooldown cooldown = new Cooldown(PD_COOLDOWN.getValue());
 
 	private int give = 0; // 주는 대미지
 	private int receive = 0; // 받는 대미지
